+++ packages/format-library/src/italic/index.js
@@ -8,6 +8,7 @@ import {
 	RichTextShortcut,
 	__unstableRichTextInputEvent,
 } from '@wordpress/block-editor';
+import { formatItalic } from '@wordpress/icons';
 
 const name = 'core/italic';
 const title = __( 'Italic' );
@@ -36,7 +37,7 @@ export const italic = {
 				/>
 				<RichTextToolbarButton
 					name="italic"
-					icon="editor-italic"
+					icon={ formatItalic }
 					title={ title }
 					onClick={ onClick }
 					isActive={ isActive }